Data Structures

QXTns has data structures for representing tensors, tensor networks and tensor network circuits. Here we go through each and show examples and features.

QXTensor

QXTns.QXTensor — Type

Datastructure representing tensors

Tensor Network

QXTns.TensorNetwork — Type

Tensor network data-structure

Tensor Network Circuit

QXTns.TensorNetworkCircuit — Type

Tensor network circuit data-structure