QXTn

QXTn is a Julia package with data structures and utilities for manipulating tensor networks. As well as generic tensor network data structure, it also contains specific data structures for handling tensor networks derived from quantum circuits. It was developed as part of the QuantEx project, one of the individual software projects of WP8 of PRACE 6IP.

It uses some features from ITensors and NDTensors for representing tensors and indices and performing contractions.

Installation

QXTn is a Julia package and can be installed using Julia's inbuilt package manager from the Julia REPL using.

import Pkg
Pkg.add("QXTn")

To ensure everything is working, the unittests can be run using

import Pkg; Pkg.test()

Example usage

An example of creating a simple tensor network and contracting.

using QXTn

tn = TensorNetwork()

a, b, c, d = Index(2), Index(3), Index(5), Index(4)

# add a 2x3x5 rank tensor
push!(tn, [a, b, c], rand(2, 3, 5))
# add a 5x4 matrix
push!(tn, [c, d], rand(5, 4))

# contract network
simple_contraction!(tn)

# number of tensors after contraction
@show length(tn)

# resulting tensor has dimensions should have dimensions 2x3x4
@show size(first(tn))

Contributing

Contributions from users are welcome and we encourage users to open issues and submit merge/pull requests for any problems or feature requests they have. The CONTRIBUTING.md on the top level of the source folder has further details of the contribution guidelines.

Building documentation

QXTn.jl uses Documenter.jl to generate documentation. To build the documentation locally run the following from the root folder.

The first time it is will be necessary to instantiate the environment to install dependencies

julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'

and then to build the documentation

julia --project=docs/ docs/make.jl

The generated document will be in the docs/build folder. To serve these locally one can use the LiveServer package as

julia --project -e 'import Pkg; Pkg.add("LiveServer");
julia --project -e  'using LiveServer; serve(dir="docs/build")'

Or with python3 using from the docs/build folder using

python3 -m http.server

The generated documentation should now be viewable locally in a browser at http://localhost:8000.

API Reference

QXTn.MockTensor

QXTn.QXTensor

QXTn.QXTensor

QXTn.QXTensor

QXTn.QXTensor

QXTn.QXTensor

QXTn.TensorNetwork

QXTn.TensorNetwork

QXTn.TensorNetworkCircuit

Base.convert

Base.convert

Base.copy

Base.delete!

Base.delete!

Base.merge

Base.push!

Base.push!

Base.push!

Base.show

Base.size

NDTensors.inds

NDTensors.store

QXTn._contract_ncon_indices

QXTn.add_input!

QXTn.add_output!

QXTn.contract_hyper_indices

QXTn.contract_ncon_indices

QXTn.contract_pair!

QXTn.contract_tensors

QXTn.contract_tn!

QXTn.create_test_tnc

QXTn.decompose_gate

QXTn.decompose_tensor!

QXTn.find_hyper_edges

QXTn.get_hyperedges

QXTn.hyperindices

QXTn.isdiagonal

QXTn.isdiagonal

QXTn.mock_contract

QXTn.neighbours

QXTn.push_input!

QXTn.push_output!

QXTn.reduce_tensor

QXTn.replace_tensor_symbol!

QXTn.replace_with_svd!

QXTn.simple_contraction

QXTn.simple_contraction!

QXTn.tensor_data

QXTn.MockTensor — Type

Tensor store struct that just tracks tensor dimensions

QXTn.QXTensor — Type

QXTensor(indices::Array{<:Index, 1},
         hyper_indices::Array{<:Array{Int64, 1}, 1},
         storage::Union{Nothing, <: NDTensors.TensorStorage}=nothing)

QXTensor constructor which creates a new instance of QXTensor with the given indices and hyper indices. If no storage data structure is given then MockTensor of that shape is added as the storage.

QXTn.QXTensor — Method

QXTensor(a::T) where T <: Number

QXTensor constructor which creates a new instance of QXTensor corresponding to a scalar

QXTn.QXTensor — Method

QXTensor(indices::Array{<:Index, 1})

QXTensor constructor to create an instance of QXTensor with the given indices and will use MockTensor for the data storage.

QXTn.QXTensor — Method

QXTensor(data::AbstractArray{Elt, N},
         indices::Array{<:Index, 1};
         diagonal_check::Bool=true) where {Elt, N}

Constructor to create a QXTensor instance using the given data and indices. If diagonalcheck is true, it will automaticallly check which indices are hyper indices and record in the hyperindices field.

QXTn.TensorNetwork — Method

TensorNetwork(array::Vector{<: QXTensor})

Outer constructor to create a tensor network object from an array of ITensor objects.

Base.convert — Method

Base.convert(::Type{ITensor}, t::QXTensor)

Overloaded convert to enable conversion of a QXTensor to an ITensor. Note that doing this will mean that information about the hyper indices will be lost.

Base.convert — Method

Base.convert(::Type{QXTensor}, t::ITensor{N}) where N

Overloaded convert to enable conversion of an ITensor to a QXTensor. Note that converting in this way means that hyper indices won't be identified.

Base.copy — Method

Overload functions from base to make MockTensor usable

Base.delete! — Method

delete!(tn::TensorNetwork, tensor_id::Symbol)

Function to remove a tensor from a tensor network.

Base.delete! — Method

delete!(tnc::TensorNetworkCircuit, tensor_id::Symbol)

Function to remove a tensor from a tensor network circuit.

Base.merge — Method

merge(a::TensorNetwork, b::TensorNetwork)

Join two networks together

Base.push! — Method

push!(tn::TensorNetwork,
      tensor::QXTensor;
      tid::Union{Nothing, Symbol}=nothing) where {N}

Function to add a tensor to the tensor network.

Keywords

tid::Union{Nothing, Symbol}=nothing: the id for the new tensor in tn. An id is

generated if one is not set.

Base.push! — Method

push!(tn::TensorNetwork,
      indices::Vector{Index},
      data::Array{T, N}
      tid::Union{Nothing, Symbol}=nothing) where {T, N}

Function to add a tensor to the tensor network.

Keywords

tid::Union{Nothing, Symbol}=nothing: the id for the new tensor in tn. An id is

generated if one is not set.

Base.push! — Method

push!(tnc::TensorNetworkCircuit,
      qubits::Vector{Int64},
      data::Array{T, 2}) where T

Function to add a gate to the tensor network circuit given the qubits it acts on and an array of the matrix elements

Base.show — Method

Custom show for QXTensors

Base.size — Method

Custom size function

NDTensors.inds — Method

Implement inds for QXTensor

NDTensors.store — Method

Implement store for QXTensor

QXTn._contract_ncon_indices — Method

_contract_ncon_indices(A_inds::IndexSet{M}, B_inds::IndexSet{N}) where {M, N}

Function return indices in ncon format for contraction of tensors with given index sets. Returns two tuples for indices in each with convention that negative values are remaining indices and positive values are indices being contracted over.

For example if (1, -1), (-2, 1) is returned, this menas that the first index of tensor A A is contracted with the second index of tensor B and the resulting tensor will have indices corresponding to the second index of the first tensor and first index of the second tensor.

QXTn.add_input! — Function

add_input!(tnc::TensorNetworkCircuit; input::Union{String, Nothing}=nothing)

Function to add input tensors to the circuit

QXTn.add_output! — Function

add_output!(tnc::TensorNetworkCircuit; output::Union{String, Nothing}=nothing)

Function to add output tensors to the circuit

QXTn.contract_hyper_indices — Method

contract_hyper_indices(a_indices::Array{<:Index, 1},
                       a_hyper_indices::Array{<:Array{<:Index, 1}, 1},
                       b_indices::Array{<:Index, 1},
                       b_hyper_indices::Array{<:Array{<:Index, 1}, 1})

This function calculates the resulting groups of hyper indices after contracting tensors with the given groups of hyper indices.

QXTn.contract_ncon_indices — Method

contract_ncon_indices(tn::TensorNetwork, A_sym::Symbol, B_sym)

Function return indices in ncon format for contraction of tensors with given symbols. Returns two tuples for indices in each with convention that negative values are remaining indices and positive values are indices being contracted over.

For example if (1, -1), (-2, 1) is returned, this menas that the first index of tensor A A is contracted with the second index of tensor B and the resulting tensor will have indices corresponding to the second index of the first tensor and first index of the second tensor.

QXTn.contract_pair! — Function

contract_pair!(tn::TensorNetwork, A_id::Symbol, B_id::Symbol; mock::Bool=false)

Contract the tensors in 'tn' with ids 'Aid' and 'Bid'. If the mock flag is true then the new tensor will be a mock tensor with the right dimensions but without the actual data.

The resulting tensor is stored in tn under the symbol C_id if one is provided, otherwise a new id is created for it.

QXTn.contract_tensors — Method

contract_tensors(A::QXTensor, B::QXTensor; mock::Bool=false)

Function to contract two QXTensors and return another QXTensor. If the mock flag is false or either of the input tensors use MockTensor then the storage for the final tensor will be of type MockTensor.

QXTn.contract_tn! — Method

contract_tn!(tn::TensorNetwork, plan)

Contract the indices of 'tn' according to 'plan'.

QXTn.create_test_tnc — Method

create_test_tnc(;input::Union{String, Nothing}=nothing,
                output::Union{String, Nothing}=nothing,
                no_input::Bool=false,
                no_output::Bool=false,
                kwargs...)

Create a tensor network circuit for a small example circuit, 3 qubit ghz preparation circuit in this case

QXTn.decompose_gate — Function

function decompose_gate!(gate_data::Array{<:Number, 4},
                         threshold::AbstractFloat=1e-15)

Function to decompose a tensor into two smaller tensors

QXTn.decompose_tensor! — Method

decompose_tensor!(tn::TensorNetwork,
                  tensor_id::Symbol,
                  left_indices::Array{<:Index, 1};
                  contract_S_with::Symbol=:V,
                  kwargs...)

Function to decompose a tensor in a tensor network using svd.

Keywords

contract_S_with::Symbol=:V: the maxtrix which should absorb the matrix of singular values

maxdim::Int: the maximum number of singular values to keep.

mindim::Int: the minimum number of singular values to keep.

cutoff::Float64: set the desired truncation error of the SVD.

QXTn.find_hyper_edges — Method

function find_hyper_edges(A::AbstractArray{Elt, N}) where {Elt, N}

Function to identify hyper edges of tensors. Returns an array of tuples of indices of the original tensor which can be identified

QXTn.get_hyperedges — Method

get_hyperedges(tn::TensorNetwork)::Array{Array{Symbol, 1}, 1}

Return an array of hyperedges in the given tensornetwork tn.

Hyperedges are represented as arrays of tensor symbols.

QXTn.hyperindices — Method

hyperindices(t::QXTensor)

Function to get the hyper indices as an array of arrays of Indices

QXTn.isdiagonal — Method

isdiagonal(A::AbstractArray{Elt, 2}) where Elt

Function to check if the given matrix is diagonal

QXTn.isdiagonal — Method

isdiagonal(A::AbstractArray{Elt, N}, Pair{Int64, Int64}) where {Elt, N}

Function to check if the given matrix is diagonal along given axes

QXTn.mock_contract — Function

mock_contract(T1::NDTensors.Tensor,
              labelsT1,
              T2::NDTensors.Tensor,
              labelsT2,
              labelsR = NDTensors.contract_labels(labelsT1, labelsT2))

Overloaded contract function from NDTensors which implements contraction for tensors using MockTensor objects as storage.

QXTn.neighbours — Method

neighbours(tn::TensorNetwork, tensor::Symbol)

Function get the symbols of the neighbouring tensors

QXTn.push_input! — Method

push_input!(tnc::TensorNetworkCircuit, tensor::Array{Elt, 1}, pos::Int64) where Elt

Function to add a single input tensor to the tensor network circuit at the given position

QXTn.push_output! — Method

push_output!(tnc::TensorNetworkCircuit, tensor::Array{Elt, 1}, pos::Int64) where Elt

Function to add a single output tensor to the tensor network circuit at the given position

QXTn.reduce_tensor — Method

reduce_tensor(A::AbstractArray{Elt, N}, hyper_index_groups::Array{Int64, 1})

Function to reduce the dimension of the given tensor assuming the given hyper edge groups. For example a diagonal matrix will have a single hyper edge group with both indices [1, 2]

julia> QXTn.reduce_tensor([[1, 0] [0, 2]], [[1, 2]])
2-element Vector{Int64}:
 1
 2

QXTn.replace_tensor_symbol! — Method

replace_tensor_symbol!(tn::TensorNetwork, orig_sym::Symbol, new_sym::Symbol)

Replace the given symbol with the given new symbol

QXTn.replace_with_svd! — Method

replace_with_svd!(tn::TensorNetwork, 
                  tensor_id::Symbol,
                  left_indices::Array{<:Index, 1};
                  kwargs...)

Function to replace a tensor in a tensor network with its svd.

The indices contained in 'left_indices' are considered the row indices of the tensor when the svd is performed.

Keywords

maxdim::Int: the maximum number of singular values to keep.

mindim::Int: the minimum number of singular values to keep.

cutoff::Float64: set the desired truncation error of the SVD.

QXTn.simple_contraction! — Method

simple_contraction!(tn::TensorNetwork)

Function to perfrom a simple contraction, contracting all tensors in order. Only useful for very small networks for testing.

QXTn.simple_contraction — Method

simple_contraction(tn::TensorNetwork)

Function to perfrom a simple contraction, contracting all tensors in order. Only useful for very small networks for testing.

QXTn.tensor_data — Method

tensor_data(tensor::QXTensor; consider_hyperindices::Bool=false)

Get the data associated with the given tensor. If the consider_hyperindices flag is true then only the first of the hyper indices are retained. For example for a 5 rank tensor where the 2nd and 4th indices form a group of hyper indices, with this option set to true would return a rank 4 tensor where the 2nd index